requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.36
openai==1.59.9
pypdf==4.3.1
python-docx==1.1.2
pydantic-settings==2.7.0
